requirements.txt
llama-cpp-python==0.1.72
pip-chill==1.0.3
sentence-transformers==2.2.2
